kill.1.adoc: clarify syntax of -SIG argument in synopsis
kill.1.adoc has this synopsis line

	*kill*  [-signal|*-s* _signal_]

For "-signal", the leading "-" should be typed literally, and the
"signal" suffix should be replaced by a signal name/number.
This means that according to GNU man(1), the "-" should render in
bold font whereas the "signal" placeholder should use italic (which
terminals often render with underscores).

Make it so. Use two askerisks because one is not enough for sub-word
markup [*].

[*] https://docs.asciidoctor.org/asciidoc/latest/text/#unconstrained

Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>