getopt.1.adoc: render synopsis rules on separate lines
AsciiDoc ignores single line breaks, so the synopsis of

	*getopt* _variant1_
	*getopt* _variant2_

will render as one line in the manpage.

Fix this by adding a blank line in between two rules, like we do
elsewhere.  This means that the manpage also has blank lines.

	*getopt* _variant1_

	*getopt* _variant2_

Alternative solution: we could avoid the blank lines like below,
but that'd be inconsistent with our other manpages.

	*getopt* _variant1_ +
	*getopt* _variant2_ +

or, equivalently, use the verse[*] directive

	[verse]
	*getopt* _variant1_
	*getopt* _variant2_

[*]: https://docs.asciidoctor.org/asciidoc/latest/blocks/verses/

Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>