unshare: Fix PDEATHSIG race for --kill-child with --pid
Exit the child explicitly should the parent terminate
just before the child invokes prctl(PR_SET_PDEATHSIG).

Note that if the child is created as pid 1 in a
new pid namespace, getppid(2) will return 0 because
the parent resides in a different namespace, and
SIGKILL will only be delivered if sent from the
ancestor namespace.

Instead of checking the pid of the parent, use pairs
of pipes to both synchronise the actions of the parent and
the child and allow each to detect the termination
of the other.

The issue can be reproduced using the following script
which must be run as root to permite the --pid option.
The script will sometimes print 0..9 despite the parent
being killed:

  while : ; do
    cat /proc/uptime
    sh -c '
      unshare --kill-child --pid --fork bash -c "
        for X in {0..9} ; do echo \$X || break ; sleep 1 ; done" &
      sleep 0 # sleep 0.00$RANDOM
      kill -9 $! 2>/dev/null' |
    cat
  done

Signed-off-by: Earl Chew <earl_chew@yahoo.com>