libfdisk: (gpt) Add UUID for Marvell Armada 3700 Boot partition
This GPT UUID is used by BootROM on Marvell Armada 3700 SoC. BootROM
searches for GPT partition with this UUID on attached SATA disk and loads
bootloader from it.

Marvell Armada 3700 SoC is used e.g. on popular EspressoBIN SBC board which
has relatively solid Linux support.

Information about SATA booting and this GPT partition is currently
documented only in development version of ARM Trusted-Firmware-A
documentation:

https://trustedfirmware-a.readthedocs.io/en/latest/plat/marvell/armada/build.html

It is missing in official documentation for Armada 3720 BootROM and
therefore hard to find it.

Signed-off-by: Pali Rohár <pali.rohar@gmail.com>