test: (lsfd) ignore noatime mnt flag when testing a fd opening / directory
Close #1717.

If the file system where the target directory is is mounted with
"noatime" flag, Linux kernel may set O_NOATIME to the fd associated
with the target directory. As a result, mkfds-directory test failed.

With this change, mkfds-directory deletes "noatime" from the output
of lsfd before verifying the output if "noatime" mount flag is set
to the file system where the directory opened by mkfds-directory is.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>