Rename: fix empty 'from' with / in filename.
Previously, if we did e.g.

    rename '' _ ./foo.txt

it would try to rename to ._/foo.txt, because `string_replace` was
passed `s="/foo.txt"` and `from` would match at the front. Now we pass
`s="foo.txt"`.

This doesn't affect the case when `to` contains a slash. In that
situation we'll still place `to` at the very beginning of the path, e.g.

    rename '' _/ ./foo.txt # rename to _/./foo.txt
    rename '' /_ ./foo.txt # rename to /_./foo.txt