lsfd: use TYPE column to show cooked file types
lsfd has two exclusive requirements：

A. show quickly understandable information in limited display area, and
B. provide primitive raw information for post-process tools.

For the purpose A, summarizing information is nice. However, the
summarized information is not suitable for tools.

To satisfy the both requirements, I plan introducing new fields
with a new convention.

STTYPE represents the type information returned from stat(2).
TYPE is the same as STTYPE but two exceptions; TYPE is filled with
PROTONAME value if the target fd is a socket, or with AINODECLASS
value if the target fd is a anonymous inode.

Below an example output:

    $ ./lsfd -o COMMAND,TYPE,STTYPE,NAME -Q '(TYPE != STTYPE)'  | head
    COMMAND                TYPE STTYPE NAME
    systemd         UNIX-STREAM   SOCK UNIX-STREAM:[25493]
    systemd         UNIX-STREAM   SOCK UNIX-STREAM:[25493]
    systemd                UNIX   SOCK UNIX:[31445]
    systemd           eventpoll   UNKN anon_inode:[eventpoll]
    systemd            signalfd   UNKN anon_inode:[signalfd]
    systemd             inotify   UNKN anon_inode:inotify
    systemd             timerfd   UNKN anon_inode:[timerfd]
    systemd           eventpoll   UNKN anon_inode:[eventpoll]
    systemd             inotify   UNKN anon_inode:inotify

STTYPE, column for tools, is printed only when the column is specified
with -o option.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>